Update set-up.md
Adding steps to get the account working after upgrading from a working test account to a paid account.  Some of this was found in the Wiki pages, but the HTTP Get on the reply path was trial and error.  I've not included images since the setup page pulls in things like phone number and api_id that are not changeable.